test: hoist MakeTestPeer into the shared net test util
llmq_chainlock_tests.cpp carried a byte-for-byte copy of the helper in
net_tests.cpp, differing only in the base address it derives the peer
address from. Move it to test/util/net, where the other net test helpers
already live, and drop both copies.

Also drop the test/util/validation.h include added alongside that copy:
nothing in llmq_chainlock_tests.cpp uses TestChainState.
